Add i2c.rs for hal.
