+++ src/i2c/ast1060_i2c.rs
@@ -1,4 +1,5 @@
 use crate::common::{DmaBuffer, DummyDelay};
+use crate::i2c::common::{I2cConfig, I2cSEvent, I2cXferMode};
 use crate::uart::UartController;
 use ast1060_pac::{I2cbuff, I2cglobal, Scu};
 use core::cmp::min;
@@ -10,31 +11,6 @@ use embedded_io::Write;
 use proposed_traits::i2c_target::I2CTarget;
 
 static I2CGLOBAL_INIT: AtomicBool = AtomicBool::new(false);
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(u32)]
-pub enum I2cSpeed {
-    Standard = 100_000,
-    Fast = 400_000,
-    FastPlus = 1_000_000,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(u8)]
-pub enum I2cXferMode {
-    DmaMode,
-    BuffMode,
-    ByteMode,
-}
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-#[repr(u8)]
-pub enum I2cSEvent {
-    SlaveRdReq,
-    SlaveWrReq,
-    SlaveRdProc,
-    SlaveWrRecvd,
-    SlaveStop,
-}
 
 pub trait Instance {
     fn ptr() -> *const ast1060_pac::i2c::RegisterBlock;
@@ -238,78 +214,6 @@ impl embedded_hal::i2c::Error for Error {
     }
 }
 
-pub struct TimingConfig {
-    pub manual_scl_high: u8,
-    pub manual_scl_low: u8,
-    pub manual_sda_hold: u8,
-    pub clk_src: u32,
-}
-pub struct I2cConfig {
-    pub xfer_mode: I2cXferMode,
-    pub multi_master: bool,
-    pub smbus_timeout: bool,
-    pub smbus_alert: bool,
-    pub timing_config: TimingConfig,
-    pub speed: I2cSpeed,
-}
-pub struct I2cConfigBuilder {
-    xfer_mode: I2cXferMode,
-    multi_master: bool,
-    smbus_timeout: bool,
-    smbus_alert: bool,
-    timing_config: Option<TimingConfig>,
-    speed: I2cSpeed,
-}
-impl I2cConfigBuilder {
-    pub fn new() -> Self {
-        Self {
-            xfer_mode: I2cXferMode::ByteMode,
-            multi_master: false,
-            smbus_alert: false,
-            smbus_timeout: false,
-            timing_config: None,
-            speed: I2cSpeed::Standard,
-        }
-    }
-    pub fn xfer_mode(mut self, mode: I2cXferMode) -> Self {
-        self.xfer_mode = mode;
-        self
-    }
-    pub fn multi_master(mut self, enabled: bool) -> Self {
-        self.multi_master = enabled;
-        self
-    }
-    pub fn smbus_alert(mut self, enabled: bool) -> Self {
-        self.smbus_alert = enabled;
-        self
-    }
-    pub fn smbus_timeout(mut self, enabled: bool) -> Self {
-        self.smbus_timeout = enabled;
-        self
-    }
-    pub fn speed(mut self, speed: I2cSpeed) -> Self {
-        self.speed = speed;
-        self
-    }
-    pub fn timing_config(mut self, config: TimingConfig) -> Self {
-        self.timing_config = Some(config);
-        self
-    }
-    pub fn build(self) -> I2cConfig {
-        I2cConfig { 
-            xfer_mode: self.xfer_mode,
-            multi_master: self.multi_master,
-            smbus_timeout: self.smbus_timeout,
-            smbus_alert: self.smbus_alert,
-            timing_config: self.timing_config.unwrap_or(TimingConfig { 
-                manual_scl_high: 0,
-                manual_scl_low: 0,
-                manual_sda_hold: 0,
-                clk_src: 0, 
-            }),
-            speed: self.speed }
-    }
-}
 
 const I2C_TOTAL: usize = 4;
 #[link_section = ".ram_nc"]
