+++ src/i2c/mod.rs
@@ -1,1 +1,3 @@
-pub mod i2ccontroller;
+pub mod common;
+pub mod ast1060_i2c;
+//pub mod i2c;
